+++ tasks/distro.js
@@ -111,12 +111,17 @@ for (const [ key, value ] of Object.entries(windowsSigningOptions)) {
   }
 }
 
-if (publish && (argv.ia32 || argv.x64 || argv.arm64)) {
+if (argv.ia32) {
+  console.error('Windows ia32 builds are no longer supported');
+  process.exit(1);
+}
+
+if (publish && (argv.x64 || argv.arm64)) {
   console.error('Do not override arch; is manually pinned');
   process.exit(1);
 }
 
-const archOptions = [ 'x64', 'ia32', 'arm64' ].filter(a => argv[a]).map(a => `--${a}`);
+const archOptions = [ 'x64', 'arm64' ].filter(a => argv[a]).map(a => `--${a}`);
 
 const extraMetadataOptions = [
   `-c.extraMetadata.SENTRY_DSN=${ process.env.SENTRY_DSN || null }`