+++ CHANGELOG.md
@@ -6,9 +6,11 @@ All notable changes to the [Camunda Modeler](https://github.com/camunda/camunda-
 
 ___Note:__ Yet to be released changes appear here._
 
+* `FEAT`: drop support for 32-bit Windows distributables
 * `FEAT`: rename cluster connection terminology to plain connection ([#6158](https://github.com/camunda/camunda-modeler/pull/6158))
 * `FIX`: do not report a connection error while the tab's connection is still being resolved
 * `FIX`: do not block drag and drop inside nested editors (e.g. reordering DMN decision table rows/columns) ([#6166](https://github.com/camunda/camunda-modeler/issue/6166))
+* `DEPS`: update to `electron@44.0.0`
 
 ## 5.51.0